+++ internal/manager/watcher_events_other.go
@@ -1,0 +1,14 @@
+//go:build !linux
+// +build !linux
+
+package manager
+
+import "github.com/syncthing/notify"
+
+func notifyEvents() []notify.Event {
+	return []notify.Event{
+		notify.Create,
+		notify.Rename,
+		notify.Write,
+	}
+}
